+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+1.0.0-dev.4
+=============
+* Fixed bugs:
+  * Fixed an issue where an exception was thrown while running functional tests in the parallel mode
+  * Fixed an issue where the getElements() method performed the search of elements in a specified context incorrectly
+  * For a typified <select> element added the ability to select strict and non-strict values
+  * Fixed an issue where an exception was thrown when running injectable tests in the parallel mode
+
 1.0.0-dev.3
 =============
 * Created Event Management and observers structure